requirements.txt
openai>=1.0.0
pyyaml>=6.0
jsonschema>=4.0.0
music21>=9.0.0
google-generativeai>=0.3.0
anthropic>=0.7.0
pandas>=1.5.0
python-dotenv>=1.0.0
